requirements.txt
chainlit
python-dotenv
openai-agents
eval_type_backport
websockets